+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased] - 2023-03-14
+- Changed: raised contao and php dependency
+- Fixed: symfony 5 compatibility
+
 ## [0.3.3] - 2022-05-09
 
 - Fixed: duplicated result item id